Update bluebirdv2
 initiates with a defined maximum supply of tokens. Upon deployment, the contract assigns tokens to a development fund, a liquidity pool, and team advisors based on predetermined rates. 

The contract includes a sliding transfer fee mechanism. This fee is activated when the amount being transferred is equal to or greater than a specified percentage of the total supply. The fee is deducted from the amount being transferred and directed to the development fund. Transfers above a certain limit are also restricted to prevent large, disruptive transfers, although certain addresses can be whitelisted to bypass this restriction. 

The contract includes a feature to interact with wsgb. A user can transact with this token, which triggers a mining operation that rewards the user with BlueBird tokens based on a set mining rate. The total mined supply is tracked, and the remaining supply is updated accordingly. To manage the mining rate, an owner-only function allows for the adjustment of the mining rate.

A mechanism for team advisors to claim their allocated tokens is also incorporated. The tokens are initially locked and become gradually claimable in quarters, with a set percentage claimable per quarter. The contract keeps track of the claimed amount for each team advisor.

The contract also has a fee contribution mechanism to the liquidity pool and the development fund for each transaction. These fees are deducted from the sender's wallet balance of wsgb and not from the total being sent.

Finally, the contract includes ReentrancyGuard protection. This security measure is crucial to prevent potential re-entrancy attacks, especially in functions that interact with external contracts or addresses. In this contract, the ReentrancyGuard is applied to functions that could potentially be targets of such attacks.